Fix duplicate Vercel assistant API route
